Fix: formatting and refactoring code
